requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==3.7.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.104.1
gitdb==4.0.11
gitpython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib-metadata==6.11.0
jinja2==3.1.4
jiter==0.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
narwhals==1.13.2
numpy==1.26.4
openai==1.54.1
packaging==23.2
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.5
psycopg2==2.9.10
psycopg2-binary==2.9.9
pyarrow==18.0.0
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-settings==2.0.3
pydeck==0.9.1
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.17
pytz==2024.2
referencing==0.35.1
requests==2.31.0
rich==13.9.4
rpds-py==0.20.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.27.0
streamlit==1.28.1
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.6
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.24.0
validators==0.34.0
watchdog==6.0.0
zipp==3.20.2
